+++ src/Components/Alfajores/Alfajores.jsx
@@ -1,17 +1,16 @@
 import React from "react";
-import Slider from "../Slider/Slider";
+import GeneralSlider from "../General_Slider/GeneralSlider";
 //CSS
 import styled from "styled-components";
-import GlobalStyles from "../../Styles/GlobalStyles";
 import { productImages } from "../../Assets";
-
+import Footer from "../Footer/Footer";
 
 
 function Alfajores() {
   return (
     <StyledContainer>
-      {/* <MainTitle>Alfajores</MainTitle> */}
-      <Slider  images={productImages} />
+      <GeneralSlider  images={productImages} />
+      <Footer/>
     </StyledContainer>
   );
 }
@@ -19,22 +18,8 @@ function Alfajores() {
 export default Alfajores;
 
 const StyledContainer = styled.div`
-  /* ${GlobalStyles.container} */
-  width: 100%;
-  height: 200vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: white;
-  margin-top: 100px;
-  /* &:hover {
-  background-color: red;
-  } */
+    width: 100%;
+    height: 100%;
 `;
-
-// const MainTitle = styled.h1`
-//   background-color: #c2c2c2;
-// `;
